Serialize randomness as a u64/BigInt to allow 2^64 search space (#1268)
* Serialize randomness as a u64/BigInt to allow 2^64 search space

This allows miners to use a bigger search space, which will allow more powerful mining rigs without them running out of search space and spending time idling. The other added benefit of this change is that it is more logical and inline with what one would expect. The previous limit of 2^53 was unexpected and could cause bugs with third party software down the road.

- Internal type is a BigInt (typescript side)
- Internal type is a u64 (rust side)
- Serialized as a big-endian hex string over network
- Pool keeps it as a hex string since it doesn't really need to do much with it

* Fixes for randomness change in DB serializer changes

* Adjust mining test

* More merge changes for DB serialization

* Remove type-casts that aren't needed anymore, fix randomness wrap-around logic

+++ ironfish/src/testUtilities/helpers/blockchain.ts
@@ -27,7 +27,7 @@ export async function makeBlockAfter(
 
   const timestamp = new Date()
   const target = Target.calculateTarget(timestamp, after.timestamp, after.target)
-  const randomness = Math.floor(Math.random() * Number.MAX_SAFE_INTEGER)
+  const randomness = BigInt(Math.floor(Math.random() * Number.MAX_SAFE_INTEGER))
   const graffiti = GraffitiUtils.fromString('fake block')
 
   const header = new BlockHeader(
